#19 reactor: Modify modal and popup component
- 주석 추가
- molecule -> atom 으로 폴더 이동
