--- src/components/molecule/modal/modal.styled.ts
@@ -1,12 +1,0 @@
-import { Box, styled } from '@mui/material';
-
-export const ModalContainer = styled(Box)(() => ({
-  backgroundColor: '#fff',
-  minWidth: 400,
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  textAlign: 'center',
-  padding: 4,
-}));
